save progress
attempt clustering ML

no avail
